Potential Neural Network (using MLP)
This is what I think the neural network should look like after watching the micrograd video. The next step would be to validate whether or not the training works using the testing dataset. If this proves to be successful, then optimization would be the next step by exploring the results of changing different parameters.